+++ src/util/functions/getDiscordUser.ts
@@ -1,12 +1,24 @@
 import Axios from "axios";
-import { User } from "discord.js";
+
+interface DiscordUser {
+	id: string;
+	username: string;
+	avatar: string;
+	discriminator: string;
+	email: string;
+	verified: boolean;
+	locale: string;
+	mfa_enabled: boolean;
+	flags: number;
+	premium_type: number;
+}
 
 export function getDiscordUser(token: string) {
-	return new Promise<User>(async (resolve, reject) => {
-		Axios("https://discord.com/api/users/@me", {
+	return new Promise<DiscordUser>(async (resolve, reject) => {
+		Axios("https://discord.com/api/v9/users/@me", {
 			headers: { Authorization: token }
 		})
-			.then(({ data }) => resolve(data as User))
+			.then(({ data }) => resolve(data as DiscordUser))
 			.catch(reject);
 	});
 }
